need to fix angles lol

+++ Bullet.py
@@ -4,7 +4,7 @@ from math import cos, sin
 
 class Bullet:
     WIDTH = 5
-    SPEED = 3
+    SPEED = 3.0
 
     def __init__(self, uid, owner, x, y, angle):
         self.uid = uid
@@ -14,11 +14,11 @@ class Bullet:
         self.angle = angle
 
     def tick(self):
-        self.x += int(cos(self.angle) * Bullet.SPEED)
-        self.y -= int(sin(self.angle) * Bullet.SPEED)
+        self.x += cos(self.angle) * Bullet.SPEED
+        self.y -= sin(self.angle) * Bullet.SPEED
 
     def getRect(self):
         return pygame.Rect(self.x - Bullet.WIDTH // 2, self.y - Bullet.WIDTH // 2, Bullet.WIDTH, Bullet.WIDTH)
 
     def state(self):
-        return ';/{0}:{1}:{2}:{3}:{4}/'.format(self.uid, self.owner, self.x, self.y, self.angle)
+        return ';/{0}:{1}:{2}:{3}:{4}/'.format(self.uid, self.owner, int(self.x), int(self.y), self.angle)
